requirements.txt
python-binance==1.0.16
numpy==1.26.0
pandas==2.1.1
TA-Lib==0.4.28
Flask==3.0.0
matplotlib==3.8.0
aiohttp==3.8.5
python-dateutil==2.8.2
python-dotenv==1.0.0
loguru==0.7.2
pytest==7.4.2
black==23.9.1
flake8==6.1.0
mypy==1.5.1
Sphinx==7.2.6
scikit-learn==0.24.2